Added modularity to the code and improved error reporting so that it's more specific and detailed
